Rename integration test prefix and helper header (reviewer feedback)
QnnInt_ prefix and qnn_int_test_utils.h filename both read as "integer"
at first glance. Align with the original 5-tier plan naming (QnnInteg_)
and drop the "int" abbreviation from the header filename.

- integration/qnn_int_test_utils.h → integration/qnn_test_utils.h
- QnnInt_OrtApiTest → QnnInteg_OrtApiTest (9 tests)
- README examples and gtest filter updated
